+++ src/mesh.cpp
@@ -22,7 +22,7 @@ void Mesh::UpdateCenterOfMass(void) {
 	centerOfMass = V3();
 	for (size_t i = 0; i < vertexCount; i++)
 		centerOfMass += vertices[i];
-	centerOfMass /= vertexCount;
+	centerOfMass /= (float) vertexCount;
 }
 
 void Mesh::Translate(const V3 &delta) {
@@ -221,9 +221,9 @@ void Mesh::DrawWireframe(Window &wind, const PPCamera &camera) {
 			camera.ProjectPoint(vertices[tri[2]], p[2])
 		) {
 
-			wind.DrawLine(p[0].x(), p[0].y(), p[1].x(), p[1].y(), colors[tri[0]], colors[tri[1]]);
-			wind.DrawLine(p[1].x(), p[1].y(), p[2].x(), p[2].y(), colors[tri[1]], colors[tri[2]]);
-			wind.DrawLine(p[0].x(), p[0].y(), p[2].x(), p[2].y(), colors[tri[0]], colors[tri[2]]);
+			wind.DrawLine((int)p[0].x(), (int)p[0].y(), (int)p[1].x(), (int)p[1].y(), colors[tri[0]], colors[tri[1]]);
+			wind.DrawLine((int)p[1].x(), (int)p[1].y(), (int)p[2].x(), (int)p[2].y(), colors[tri[1]], colors[tri[2]]);
+			wind.DrawLine((int)p[0].x(), (int)p[0].y(), (int)p[2].x(), (int)p[2].y(), colors[tri[0]], colors[tri[2]]);
 
 		}
 	}
